feat(Peer Group): Add ability to get manually assigned peer group
If a Peer Activity uses auto assignment and a Peer Group is requested
and the Peer Group does not exist, it will automatically be created.

If a Peer Activity uses manual assignment and a Peer Group is requested
and the Peer Group does not exist, it will not be automatically created.

#74